refactor(web-chat): import sessionstream from pkg/sessionstream
