+++ cmd/web-chat/agentmode_chat_feature_test.go
@@ -7,7 +7,7 @@ import (
 	gepevents "github.com/go-go-golems/geppetto/pkg/events"
 	chatapp "github.com/go-go-golems/pinocchio/pkg/chatapp"
 	agentmode "github.com/go-go-golems/pinocchio/pkg/middlewares/agentmode"
-	sessionstream "github.com/go-go-golems/sessionstream"
+	sessionstream "github.com/go-go-golems/sessionstream/pkg/sessionstream"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/types/known/structpb"
 )
