+++ cmd/web-chat/app/server.go
@@ -14,10 +14,10 @@ import (
 
 	chatapp "github.com/go-go-golems/pinocchio/pkg/chatapp"
 	infruntime "github.com/go-go-golems/pinocchio/pkg/inference/runtime"
-	sessionstream "github.com/go-go-golems/sessionstream"
-	storememory "github.com/go-go-golems/sessionstream/hydration/memory"
-	storesqlite "github.com/go-go-golems/sessionstream/hydration/sqlite"
-	wstransport "github.com/go-go-golems/sessionstream/transport/ws"
+	sessionstream "github.com/go-go-golems/sessionstream/pkg/sessionstream"
+	storememory "github.com/go-go-golems/sessionstream/pkg/sessionstream/hydration/memory"
+	storesqlite "github.com/go-go-golems/sessionstream/pkg/sessionstream/hydration/sqlite"
+	wstransport "github.com/go-go-golems/sessionstream/pkg/sessionstream/transport/ws"
 	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/proto"
 )
