+++ pkg/chatapp/service_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	sessionstream "github.com/go-go-golems/sessionstream"
+	sessionstream "github.com/go-go-golems/sessionstream/pkg/sessionstream"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/types/known/structpb"
 )
